+++ src/lib/tools.ts
@@ -1,10 +1,13 @@
 import { URL } from "url"
-import { RocketChatStatus } from ".."
+import { RCSNext } from "../next"
 
-export function buildCommand(cmd: RocketChatStatus.Command): string {
+export function buildCommand(cmd: RCSNext.Base.Command): string {
     return `rocket-chat-status.${cmd}`
 }
 
+/**
+ * @deprecated
+ */
 export function validateHost(url: string): boolean {
     try {
         new URL(url)